feat: fix issues with website demo
