+++ packages/vstory/demo/src/demos/VChartSite/scene11.tsx
@@ -289,7 +289,7 @@ export const scene11Characters: ICharacterSpec[] = [
     }
   },
   {
-    type: 'WordCloudChart',
+    type: 'VChart',
     id: `scene11-wordcloud-chart`,
     zIndex: 0,
     position: {
