+++ packages/vstory/demo/src/demos/VChartSite/scene9.tsx
@@ -260,7 +260,7 @@ export const scene9Characters: ICharacterSpec[] = [
     }
   },
   {
-    type: 'RoseChart',
+    type: 'VChart',
     id: `scene9-rose-chart`,
     zIndex: 2,
     position: {
@@ -327,6 +327,7 @@ export const scene9: ISceneSpec = {
             animation: {
               duration: 500,
               easing: easeInOutQuad,
+              effect: 'move',
               move: {
                 from: 'right'
               }
@@ -421,18 +422,18 @@ export const scene9: ISceneSpec = {
             }
           }
         },
-        ...month.map((mon, i) => {
-          return {
-            action: 'add',
-            startTime: 2000 + (i + 1) * 200,
-            duration: 200,
-            payload: {
-              id: '1856Rose',
-              // @ts-ignore
-              values: monthData[mon]
-            }
-          };
-        }),
+        // ...month.map((mon, i) => {
+        //   return {
+        //     action: 'add',
+        //     startTime: 2000 + (i + 1) * 200,
+        //     duration: 200,
+        //     payload: {
+        //       id: '1856Rose',
+        //       // @ts-ignore
+        //       values: monthData[mon]
+        //     }
+        //   };
+        // }),
         {
           action: 'disappear',
           startTime: 5500,
@@ -463,6 +464,7 @@ scene9.actions.forEach(({ characterId, characterActions }) => {
         animation: {
           duration: 500,
           easing: easeInOutQuad,
+          effect: 'move',
           move: {
             to: 'left'
           }
