+++ packages/vstory/src/player/processor/chart/component/axisAppear.ts
@@ -1,5 +1,6 @@
 import type { IChartVisibilityAction } from '../../interface/appear-action';
-import { commonFade, commonGrow } from './commonAppear';
+import { commonFade } from '../common/commonAppear';
+import { commonGrow } from './commonAppear';
 import type { IGroup, ILine, IText } from '@visactor/vrender-core';
 
 // TODO: 区分直角坐标系和极坐标系
